Factored FormEntryAccessPreferences out to itss own class
